Add files via upload
This was an excel project done in Spring 2017 at NYU London. The Project covered in a first part portfolio theory using data from a variety of stocks and trying to build an optimal portfolio of 5 assets using portfolio optimization methods. The second part of the project consisted of using the Capital Asset Pricing Model (CAPM) to calculate market and firm-specific risk for the five assets.